[test] Cover http/tcp/icmp/browser under OTel heartbeat PL e2e
Expand the Fleet private-location OTel suite so every synthetics monitor
type asserts a HEALTHY beats-receiver component, keeping browser as the
#15968 schedule-stream regression case.
